requirements.txt
aiocache==0.11.1
aiocqhttp==1.4.3
aiodns==3.0.0
aiofiles==22.1.0
aiohttp==3.8.1
aiomysql==0.1.1
aiosignal==1.2.0
aiosqlite==0.17.0
altgraph==0.17.3
amis-python==1.0.7
anyio==3.6.1
appdirs==1.4.4
APScheduler==3.9.1
arrow==1.2.2
asgiref==3.5.2
astroid==1.6.6
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
atlastk==0.13.2
attrs==21.4.0
autopep8==1.3.5
Babel==2.12.1
bcrypt==4.0.1
beautifulsoup4==4.11.1
binaryornot==0.4.4
blinker==1.4
Brotli==1.0.9
bs4==0.0.1
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
channels==4.0.0
chardet==4.0.0
charset-normalizer==2.0.12
chevron==0.14.0
click==8.1.3
colorama==0.4.4
config==0.5.1
cookiecutter==1.7.3
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.1.0
cycler==0.11.0
databases==0.5.5
decorator==5.1.0
distlib==0.3.6
dnspython==2.2.1
docxcompose==1.4.0
docxtpl==0.16.6
drf-yasg==1.9.0
dwebsocket==0.5.12
ecdsa==0.18.0
email-validator==1.3.0
entrypoints==0.3
et-xmlfile==1.1.0
fastapi==0.73.0
fastapi_amis_admin==0.4.0
fastapi_user_auth==0.4.1
filelock==3.8.2
flex==6.13.2
fonttools==4.34.4
frozenlist==1.3.0
future==0.18.2
fuzzyfinder==2.1.0
gitdb==4.0.9
GitPython==3.1.29
greenlet==1.1.2
gunicorn==19.8.1
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpie==0.9.9
httptools==0.4.0
httpx==0.23.0
httpx-socks==0.7.3
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.11.4
inflection==0.5.1
ipykernel==4.10.1
ipython==4.0.3
iso8601==1.1.0
isort==5.10.1
itsdangerous==2.1.2
itypes==1.2.0
jieba==0.42.1
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.1.0
jsonpointer==1.14
jsonschema==4.17.3
jupyter-client==7.1.0
jupyter-core==4.9.1
kiwisolver==1.4.4
lazy-object-proxy==1.8.0
loguru==0.6.0
logzero==1.7.0
lxml==4.9.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
mccabe==0.7.0
MouseInfo==0.1.3
msgpack==1.0.4
multidict==6.0.2
mysql-connector-python==8.0.29
Naked==0.1.32
nb-cli==0.6.6
NeedForCryptography==1.5
nest-asyncio==1.5.1
nonebot-adapter-onebot==2.1.5
nonebot-plugin-gocqhttp==0.5.4
nonebot2==2.0.0rc3
numpy==1.23.4
openai==0.27.0
openapi-codec==1.3.2
opencv-python==4.7.0.68
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==2.0.1
passlib==1.7.4
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
playwright==1.22.0
poyo==0.5.0
priority==2.0.0
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
PyAutoGUI==0.9.50
pycares==4.1.2
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.5
pyee==8.1.0
pyfiglet==0.8.post1
PyGetWindow==0.0.9
Pygments==2.12.0
pygtrie==2.4.2
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
PyJWT==1.6.4
pylint==1.9.2
pymdown-extensions==9.4
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyopencl==2022.2.4
pyotp==2.2.6
pyparsing==3.0.9
pyperclip==1.8.2
pypika-tortoise==0.1.6
pypinyin==0.47.1
pyppeteer==1.0.2
pyquery==1.4.3
PyRect==0.2.0
pyrsistent==0.19.2
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-markdown-math==0.8
python-multipart==0.0.5
python-slugify==6.1.2
python-socks==2.0.3
pytools==2022.1.13
pytweening==1.0.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywebio==1.7.1
pywin32==302
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==22.3.0
quart==0.16.3
redis==4.4.0
requests==2.27.1
rfc3986==1.5.0
rfc3987==1.3.8
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.1.3
scipy==1.9.3
selenium==4.4.3
Shapely==1.8.5.post1
shellescape==3.8.1
simplegeneric==0.8.1
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a29
sqlalchemy_database==0.1.0
sqlmodel==0.0.8
sqlmodelx==0.0.4
sqlparse==0.4.2
SSIM-PIL==1.0.14
starlette==0.17.1
strict-rfc3339==0.7
swagger-spec-validator==2.1.0
tencentcloud-sdk-python==3.0.639
text-unidecode==1.3
threadpoolctl==3.1.0
toml==0.10.2
tomlkit==0.11.6
tornado==6.2
tortoise==0.1.1
tortoise-orm==0.19.3
tqdm==4.64.0
traitlets==5.1.1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
ua-parser==0.16.1
ujson==5.4.0
uritemplate==4.1.1
urllib3==1.26.9
user-agents==2.2.0
uvicorn==0.17.6
validate-email==1.3
virtualenv==20.17.1
watchfiles==0.18.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.1
Werkzeug==2.1.2
win32-setctime==1.1.0
wordcloud==1.8.2.2
wrapt==1.14.1
wsproto==1.1.0
xlrd==1.2.0
yarl==1.7.2
zipp==3.8.0
